[refactor] all the classes earlier in odemis.gui.model.init.py have been split and classified.
All constants are placed in _constants.py
FileInfo class in file_info.py
All classes with MainGUIData as base class in main_gui_data.py
All classes with StreamView as base class in stream_view.py
All classes with MicroscopyGUIData as base class in tab_gui_data.py
