+++ src/odemis/gui/model/file_info.py
@@ -1,0 +1,85 @@
+# -*- coding: utf-8 -*-
+"""
+:created: 16 Feb 2012
+:author: Éric Piel
+:copyright: © 2012 - 2022 Éric Piel, Rinze de Laat, Philip Winkler, Delmic
+
+This file is part of Odemis.
+
+.. license::
+    Odemis is free software: you can redistribute it and/or modify it under the
+    terms of the GNU General Public License version 2 as published by the Free
+    Software Foundation.
+
+    Odemis is distributed in the hope that it will be useful, but WITHOUT ANY
+    WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+    FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+    details.
+
+    You should have received a copy of the GNU General Public License along with
+    Odemis. If not, see http://www.gnu.org/licenses/.
+
+"""
+import os
+
+from odemis import model
+
+
+class FileInfo(object):
+    """
+    Represent all the information about a microscope acquisition recorded
+    inside a file. It's mostly aimed at containing information, and its
+    attributes should be considered readonly after initialisation.
+    """
+
+    def __init__(self, a_file=None, metadata=None):
+        """
+        :param a_file: (str or File or None): the full name of the file or
+            a File that contains the acquisition. If provided (and the file
+            exists), some fields will be automatically filled in.
+        :param metadata: (dict str -> value): The meta-data as model.MD_*.
+        """
+
+        self.file_name = None
+        self.file_obj = None
+
+        if isinstance(a_file, str):
+            # The given parameter is a file name
+            self.file_name = a_file
+        elif a_file is not None:
+            # Assume the given parameter is a File Object
+            self.file_name = a_file.name
+            self.file_obj = a_file # file object
+
+        # Ensure the file name contains the full path
+        self.file_name = os.path.abspath(self.file_name)
+
+        # TODO: settings of the instruments for the acquisition?
+        # Might be per stream
+        self.metadata = metadata or {}
+
+        if model.MD_ACQ_DATE not in self.metadata and self.file_name:
+            # try to auto fill acquisition time (seconds from epoch)
+            try:
+                acq_date = os.stat(self.file_name).st_ctime
+                self.metadata[model.MD_ACQ_DATE] = acq_date
+            except OSError:
+                # can't open the file => just cannot guess the time
+                pass
+
+    @property
+    def file_path(self):
+        """ Return the directory that contains the file """
+        return os.path.dirname(self.file_name) if self.file_name else None
+
+    @property
+    def file_basename(self):
+        """ Return the file name """
+        return os.path.basename(self.file_name) if self.file_name else None
+
+    @property
+    def is_empty(self):
+        return self.file_name is None
+
+    def __repr__(self):
+        return "%s (%s)" % (self.__class__, self.file_name)
